lgc/custom_tags: override django's truncatewords
lgc_truncatewords will do the same as truncatewords but will not
add any trailing characters such as an ellipsis.

The main issue with truncatewords is that its behaviour is different
depends based on its version. For the Django version 2.2.2 it adds
an ellipsis at the end and for the version 2.1.7 it adds '...'.

Signed-off-by: Krzysztof Witek <kwitek@kw-net.com>