users/logout: improve session expiry
The main problem is that when wrong crendentials are passed,
the site show up a session expiry error message.
This is because there is no track of the actual user logout
after their session has expired.
From now on we'll check the session expiration in the middleware
so we know exactly what's going on.

Signed-off-by: Krzysztof Witek <kwitek@kw-net.com>